dummy rhythms_memory working, still buggy
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@4279 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ externals/frankenstein/common.h
@@ -19,8 +19,8 @@ typedef struct t_note
 };
 typedef struct t_duration
 {
-	int numerator; // like in music notation: in a 1/4 note the numerator is 1
-	int denominator; // like in music notation: in a 1/4 note the denominator is 4
+	unsigned short int numerator; // like in music notation: in a 1/4 note the numerator is 1
+	unsigned short int denominator; // like in music notation: in a 1/4 note the denominator is 4
 };
 struct t_note_event
 {
